requirements.txt
pywin32>=306
requests>=2.31.0
customtkinter>=5.2.0
pystray>=0.19.5
Pillow>=10.0.0
